Rename bmaas_refresh_available_resources bool in NewProviderFunc
We've renamed it to bmaas_resources_available, since the intention is
that the boolean will now decide if a bmaas client is created by
terraform-provider-hpegl.  This is to avoid spurious error messages
about the lack of a .gltform file when the tenant being used doesn't
have any bmaas resources and thus the bmaas code will not be invoked.