Add support for token generation using service-client creds
In this PR we add support for generating an IAM token using
service-client creds.  The token code is held in pkg/token, which has
three sub-packages:
- common: which defines a couple of common constants, a Result struct
  and an interface that different token handlers must implement
- retrieve: which contains code to set-up a token retrieve function for
  use in provider code, which executes the interface from above to get
  the two channels that it needs
- serviceclient: which implements a token handler that uses
  service-client creds to get a token.  The token is presented to
  provider code on a channel by a go-routine that will refresh the token
  if it expires.

We've changed the gltform package so that WriteGLConfigLocally no longer
takes a token as input.  We've added omitempty to the Token field.

We've also added a couple of new configuration variables to provider.go.
Finally we've updated the README.