Change service config block to TypeList with MaxItems of 1
Currently the Terraform SDK doesn't support the use of TypeSet with
elements that have nested configuration blocks:
https://github.com/hashicorp/terraform-plugin-sdk/issues/616

The workaround is to instead use a TypeList with a MaxItems of 1.  We've
changed convertToTypeSet to convertToTypeList and made this change.
We've also had to modify client.GetServiceSettingsMap to expect a
[]interface{} as opposed to *schema.TypeSet

We've added a TODO about this so that when the SDK supports nested
resources in TypeSet then we can revert to TypeSet.